put an argument so we can see if the site was redirected
I want to be able to put a banner on the website if users come to us
via howdykolache.com. In order to do that, I need to know if on loading
howdybreakfastbuns.com it was loaded directly or sent from howdykolache.

the document.referrer seems to be empty, so manually appending ?r=1 to
check when we load howdybreakfastbuns.com